deleted models and migrations, added new migrations for the ones I deleted with corrected field names

+++ db/migrate/20190503030629_create_mentors.rb
@@ -1,7 +1,7 @@
 class CreateMentors < ActiveRecord::Migration[5.2]
   def change
     create_table :mentors do |t|
-      t.float :rate
+      t.integer :rate
       t.text :about_me
       t.references :user, foreign_key: true
